uncleaned & untested final version
