+++ auction-analysis/data/raw/flux.csv
@@ -1,0 +1,32 @@
+,Year,Open,Close
+1,1991,48,61
+2,1992,30,40
+3,1993,169,38
+4,1994,132,51
+5,1995,173,92
+6,1996,121,117
+7,1997,241,95
+8,1998,110,112
+9,1999,101,124
+10,2000,67,87
+11,2001,109,130
+12,2002,170,140
+13,2003,57,101
+14,2004,82,83
+15,2005,57,63
+16,2006,37,51
+17,2007,24,34
+18,2008,28,27
+19,2009,29,222
+20,2010,26,61
+21,2011,47,64
+22,2012,18,51
+23,2013,18,100
+24,2014,9,56
+25,2015,18,68
+26,2016,18,30
+27,2017,13,48
+28,2018,14,26
+29,2019,12,59
+30,2020,15,15
+31,2021,2,11
