+++ auction-analysis/requirements.txt
@@ -1,0 +1,6 @@
+streamlit==1.30.0
+pandas==1.5.2
+numpy==1.21.5
+matplotlib==3.6.2
+plotly==5.18.0
+geopandas==0.13.2
